Separated test file for asserting types from expecting rule errors
