+++ tests/PHPStan/Generics/GenericsIntegrationTest.php
@@ -9,6 +9,7 @@ class GenericsIntegrationTest extends \PHPStan\Testing\LevelsTestCase
 	{
 		return [
 			['functions'],
+			['functionsAssertType'],
 			['invalidReturn'],
 			['pick'],
 			['varyingAcceptor'],
