+++ tests/PHPStan/Generics/data/functions-7.json
@@ -1,7 +1,7 @@
 [
     {
         "message": "Parameter #2 $b of function PHPStan\\Generics\\Functions\\f expects callable(int): string, (callable(int): string)|null given.",
-        "line": 138,
+        "line": 37,
         "ignorable": true
     }
 ]
