+++ tests/PHPStan/Generics/data/functionsAssertType-0.json
