+++ tests/PHPStan/Generics/data/functionsAssertType-5.json
@@ -1,0 +1,12 @@
+[
+    {
+        "message": "Parameter #2 $b of function PHPStan\\Generics\\FunctionsAssertType\\f expects callable(int): mixed, null given.",
+        "line": 139,
+        "ignorable": true
+    },
+    {
+        "message": "Parameter #2 $b of function PHPStan\\Generics\\FunctionsAssertType\\f expects callable(int): mixed, '' given.",
+        "line": 140,
+        "ignorable": true
+    }
+]
