+++ tests/PHPStan/Generics/data/functionsAssertType-7.json
@@ -1,0 +1,7 @@
+[
+    {
+        "message": "Parameter #2 $b of function PHPStan\\Generics\\FunctionsAssertType\\f expects callable(int): string, (callable(int): string)|null given.",
+        "line": 138,
+        "ignorable": true
+    }
+]
